Add `history` and `revert_to` command for chat-cli; Move `message.rs` into `core` lib.

+++ chat/chat-server/src/main.rs
@@ -15,7 +15,7 @@ use rocket::{
     State,
 };
 
-type ModelParamType = half::bf16;
+type ModelParamType = f32;
 
 #[derive(Parser, Debug)]
 #[command(version)]
